// src/pages/Popup/routing/types.ts
export enum AVAILABLE_POPUP_ROUTES {
  BREAKS = '/breaks',
  SOUNDS = '/sounds',
  COLORS = '/colors',
  ACCOUNT = '/account',
  ANALYTICS = '/analytics',
  SPLASH = '/splash',
  LOADING = '/',
  WELCOME = '/welcome',
}

export enum AVAILABLE_MORE_BREAK_ROUTES {
  MOVEMENT = '/options.html#/',
  MEDITATION = '/options.html#/',
  BREATHWORK = '/options.html#/',
}

export enum AVAILABLE_CONTACT_ROUTES {
  CONTACTUS = '/contact-us',
  FAQ = '/faq',
  TERMCONDITION = '/terms-condition',
  PRIVACYPOLICY = '/privacy',
}